only link to eza theme if current theme supports it

+++ install/config/theme.sh
@@ -23,4 +23,7 @@ mkdir -p ~/.config/mako
 ln -snf ~/.config/omarchy/current/theme/mako.ini ~/.config/mako/config
 
 mkdir -p ~/.config/eza
-ln -snf ~/.config/omarchy/current/theme/eza-theme.yml ~/.config/eza/theme.yml
+# Try to link eza theme, not all themes support eza
+if [ -f ~/.config/omarchy/current/theme/eza-theme.yml ]; then
+  ln -snf ~/.config/omarchy/current/theme/eza-theme.yml ~/.config/eza/theme.yml
+fi
